requirements.txt
python-openstacksdk
